requirements.txt
pyyaml
yamale
pyperclip
